♻️ Extracted QuickCrafter to its own component file
