+++ package.json
@@ -55,10 +55,10 @@
     "html-entities": "^2.3.3",
     "json5": "^2.2.1",
     "limiter": "^2.1.0",
-    "react": "^18.2.0",
-    "react-dom": "^18.2.0",
     "mocha": "^10.0.0",
     "prettier": "^2.7.1",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
     "rollup": "^2.75.3",
     "rollup-plugin-scss": "^3",
     "rollup-plugin-userscript-metablock": "^0.3.1",
